refactor: split 3 more service files (1200-1300 LOC each)
Split 3 service files using mixin pattern:

| File | Before | After (shim) | Modules |
|---|---|---|---|
| reporting_service.py | 1274 | 6 | 3 (reporting_svc/) |
| patient_onboarding_service.py | 1256 | 4 | 4 (patient_onboarding/) |
| user_management_service.py | 1235 | 8 | 3 (user_mgmt/) |
| **Total** | **3765** | **18** | **10** |

All public APIs preserved: ReportingService, get_reporting_service,
PatientOnboardingService, UserManagementService, get_user_management_service,
user_management_service (global instance).
All 34 tests pass. API router: 1128 routes, 0 duplicates.
